### 2025-07-12 #250 @SimonMolinsky
pyinterpolate
Submitting Author: (@SimonMolinsky )
All current maintainers: (@SimonMolinsky )
Package Name: Pyinterpolate
One-Line Description of Package: Spatial interpolation: Kriging and Variogram Analysis toolset
Repository Link: https://github.com/DataverseLabs/pyinterpolate
Version submitted: 1.0.1
Archive: https://doi.org/10.5281/zenodo.17691515
JOSS DOI: https://doi.org/10.21105/joss.02869
Version accepted: 1.2.0

---

Pyinterpolate is the Python library for spatial statistics. The package provides access to spatial statistics tools (variogram analysis, Kriging, Poisson Kriging, Indicator Kriging, Universal Kriging, Inverse Distance Weighting).

Package is designed for: GIS experts, Geologists, Social scientists

The core functionalities of Pyinterpolate are spatial interpolation and spatial prediction for point and block datasets, and areal data disaggregation.

Pyinterpolate performs:

- Ordinary Kriging and Simple Kriging - spatial interpolation from points
- Centroid-based Poisson Kriging of polygons - spatial interpolation from blocks and regions
- Area-to-area and Area-to-point Poisson Kriging of Polygons - spatial interpolation and data deconvolution from areas to points
- Indicator Kriging - kriging based on probabilities
- Universal Kriging - kriging with trend
- Inverse Distance Weighting - benchmarking spatial interpolation technique
- Semivariogram regularization and deconvolution - transforming variogram of areal data in regards to point support data
- Semivariogram modeling and analysis - is your data spatially correlated? How do neighbors influence each other?

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

The target audience of `pyinterpolate` are researchers using spatial interpolation techniques in their studies (in mining, geology, social sciences, ecology, public health). Emphasis on Area-to-Point Poisson Kriging makes `pyinterpolate` valuable especially for people working with population-based spatial data that need to transform areal aggregates into point-support models.

**PyKrige** - point kriging, supports 2D and 3D ordinary and universal kriging and variogram analysis. Users cannot perform areal-based kriging. Package seems to be not maitained in the last year.

**Tobler (PySAL)** - areal interpolation, uses different modeling techniques than `pyinterpolate`. Package is under active development.

n/a
