+++ README.md
@@ -2,7 +2,7 @@
 
 # Pyinterpolate
 
-**version 1.0.2**
+**version 1.0.3**
 
 ![Logo](pyinterpolate-banner.png)
 
