### 2025-07-20 #250 @eliotwrobson
---
---

Overall, the package looks really good! I think the documentation is very clear and easy to navigate. Since all of the pre-review checks look good, I'll work on finding an editor. In the meantime, please take a look at the [readme guidelines](https://www.pyopensci.org/python-package-guide/documentation/repository-files/readme-file-best-practices.html). I don't think it's strictly necessary to address all of these before starting a review, but getting through as many items as you can while we're getting the ball rolling will help speed up the review process. Thanks again @SimonMolinsky !
