### 2025-09-10 #250 @banesullivan
:wave: Hi @crhea93 and @martinfleis! Thank you for volunteering to review pyinterpolate for pyOpenSci! I'm excited to kick off this review as editor and I'll try to chime in with a few comments of my own once I have a moment to experiment with pyinterpolate myself.

- [x] @crhea93 survey completed.
- [ ] @martinfleis survey completed.

Please get in touch with any questions or concerns!

### Your review is due: October ~3rd, 2025

I've set the due date ~3 weeks out from now, but we're flexible here so please don't rush yourself if life or other things pop up and require you to delay a bit.

---

Reviewers: @crhea93 @martinfleis
Due date: 2025-10-03
