### 2025-09-10 #250 @crhea93
- [x] As the reviewer, I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience in the README file.
- [x] **Short quickstart tutorials** demonstrating significant functionality that successfully runs locally.
- [x] **Metadata** including author(s), author e-mail(s), a URL, and any other relevant metadata, for example, in a `pyproject.toml` file or elsewhere.

*NOTE: If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be wider than high. A badge for pyOpenSci peer review will be provided when the package is accepted.*

The package structure should follow the general community best practices. In general, please consider whether:

---

I've put my comments (minor as they are) as individual GitHub issues.

Overall, this is an extremely well documented package. I congratulate the authors on their work!
