### 2025-09-12 4fbe635
[del] removed triangular selection in directional variogram methods
