+++ src/pyinterpolate/semivariogram/experimental/classes/experimental_variogram.py
@@ -50,13 +50,6 @@ class ExperimentalVariogram:
         * The baseline point is at a center of the ellipse.
         * The ``tolerance == 1`` creates an omnidirectional semivariogram.
 
-    dir_neighbors_selection_method : str, default = 't'
-        Neighbors selection in within a given angle. Available methods:
-
-        * "triangle" or "t", default method, where point neighbors are
-          selected from a triangular area,
-        * "ellipse" or "e", the more accurate method but also the slowest.
-
     custom_bins : numpy array, optional
         Custom bins for semivariance calculation. If provided, then parameter
         ``step_size`` is ignored and ``max_range`` is set to the final bin
@@ -130,7 +123,6 @@ class ExperimentalVariogram:
                  max_range: float = None,
                  direction: float = None,
                  tolerance: float = None,
-                 dir_neighbors_selection_method: str = 't',
                  custom_bins: Union[np.ndarray, Collection] = None,
                  custom_weights: np.ndarray = None,
                  is_semivariance=True,
@@ -159,7 +151,6 @@ class ExperimentalVariogram:
         self.custom_weights = custom_weights
         self.direction = direction
         self.tolerance = tolerance
-        self.method = dir_neighbors_selection_method
         self.as_cloud = as_cloud
         self.__c_sem = is_semivariance
         self.__c_cov = is_covariance
@@ -249,7 +240,6 @@ class ExperimentalVariogram:
             max_range=self.max_range,
             direction=self.direction,
             tolerance=self.tolerance,
-            dir_neighbors_selection_method=self.method,
             custom_bins=self.lags
         )
 
@@ -274,7 +264,6 @@ class ExperimentalVariogram:
             max_range=self.max_range,
             direction=self.direction,
             tolerance=self.tolerance,
-            dir_neighbors_selection_method=self.method,
             custom_bins=self.lags,
             custom_weights=self.custom_weights
         )
@@ -293,7 +282,6 @@ class ExperimentalVariogram:
             max_range=self.max_range,
             direction=self.direction,
             tolerance=self.tolerance,
-            dir_neighbors_selection_method=self.method,
             custom_bins=self.lags,
             custom_weights=self.custom_weights
         )
@@ -347,7 +335,6 @@ class ExperimentalVariogram:
                         f'max_range={self.max_range},'
                         f'direction={self.direction},'
                         f'tolerance={self.tolerance},'
-                        f'dir_neighbors_selection_method={self.method},'
                         f'custom_bins={self.lags.tolist()},'
                         f'custom_weights={custom_weights},'
                         f'is_semivariance={self.__c_sem},'
@@ -407,7 +394,6 @@ def build_experimental_variogram(ds: np.ndarray,
                                  max_range: float = None,
                                  direction: float = None,
                                  tolerance: float = None,
-                                 dir_neighbors_selection_method: str = 't',
                                  custom_bins: Union[
                                      np.ndarray, Collection
                                  ] = None,
@@ -450,13 +436,6 @@ def build_experimental_variogram(ds: np.ndarray,
         * The baseline point is at a center of the ellipse.
         * The ``tolerance == 1`` creates an omnidirectional semivariogram.
 
-    dir_neighbors_selection_method : str, default = 't'
-        Neighbors selection in within a given angle. Available methods:
-
-        * "triangle" or "t", default method, where point neighbors are
-          selected from a triangular area,
-        * "ellipse" or "e", the more accurate method but also the slowest.
-
     custom_bins : numpy array, optional
         Custom bins for semivariance calculation. If provided, then parameter
         ``step_size`` is ignored and ``max_range`` is set to the final bin
[…]
