### 2025-09-12 DataverseLabs/pyinterpolate#474 @crhea93
Update use cases to  include links
It would be very useful if the use cases page (https://pyinterpolate.readthedocs.io/en/stable/community/community/use_cases.html) had links to some of the use cases so that it is easier to understand the use of this package in real world problems.

This issue is part of the PyOpenSci review at https://github.com/pyOpenSci/software-submission/issues/250
