### 2025-09-15 DataverseLabs/pyinterpolate#474 @SimonMolinsky
Ok, I will add links where I can, and maybe section where Poisson Kriging is applied!
