### 2025-09-16 2c020e9
Updated tutorials and DirectionalVariogram class (removed triangular selection method)
