### 2025-09-16 a523d61
Removed unused custom_weights from the semivariogram point cloud function and class
