### 2025-09-16 429d114
Updated missing DOI in IndicatorVariogram class
