### 2025-09-19 3b796f0
[enh] updated Blocks geometry sampling (representative point from the largest polygon in multipolygon)
