+++ docs/source/conf.py
@@ -22,7 +22,7 @@ sys.path.insert(0, os.path.abspath('../../src'))
 project = 'pyinterpolate'
 copyright = '2025, Szymon Moliński'
 author = 'Szymon Moliński'
-release = '1.0.3'
+release = '1.1.0'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
