+++ docs/source/community/community/use_cases.rst
@@ -1,9 +1,7 @@
 Use Cases
 =========
 
-* Tick-Borne Disease Detector for the European Space Agency (2019-2020).
-* B2C project related to the prediction of demand for specific flu medications (2020).
-* B2G project related to the large-scale infrastructure maintenance (2020-2021).
-* E-commerce service for reporting and analysis, building spatial / temporal profiles of customers (2022+).
-* The external data augmentation for e-commerce services (2022+).
-* As a module for the hotel price optimization system (2023+).
+* `Lyme Disease Risk Assessment Based on the Earth Observation Data and Spatial Processing Technique. ESA Φ-week 2021 <https://phiweek.esa.int/detailed-programme>`_
+* `Disaggregation of socio-economic data <https://ml-gis-service.com/index.php/2022/12/09/get-more-from-crime-rate-data-and-other-socio-economic-indicators-with-pyinterpolate/>`_
+* `Interpolation of air pollution <https://ml-gis-service.com/index.php/2021/10/31/data-science-interpolate-air-quality-measurements-with-python/>`_
+* `Educational - Spatial Data Science for Social Geography Course - Interpolation <https://github.com/martinfleis/sds?tab=readme-ov-file#spatial-data-science-for-social-geography>`_
