### 2025-09-19 DataverseLabs/pyinterpolate#474 @SimonMolinsky
@crhea93
Is this what you meant? - https://github.com/DataverseLabs/pyinterpolate/tree/474-update-use-cases-to-include-links?tab=readme-ov-file#use-cases

I've updated use cases in README and docs, but I want to be sure that it is enough before I push it into `main` branch. Anyway, most of them are not "visible" because they have not been published in journals. I partially covered scientific research areas where Poisson Kriging is applied in `Bibliography` chapter (and JOSS paper).

In addition to this simple list I can prepare tutorial showing disaggregation of US Census data into H3 grid. What do you think about that?
