### 2025-09-19 DataverseLabs/pyinterpolate#474 @crhea93
Yes this is perfect! Thank you
