### 2025-09-19 DataverseLabs/pyinterpolate#484 @crhea93
Include CI and test coverage
As per the review template for the PyOpenSci submission, please include CI and test coverage badges. It is also useful to have codecov, but this is not a requirement.

This issue is part of the PyOpenSci review at https://github.com/pyOpenSci/software-submission/issues/250
