### 2025-09-20 9700aa8
add tests cov workflow
