+++ .github/workflows/python-install-and-test-on-linux-always.yml
@@ -4,20 +4,7 @@ on: [push]
 
 jobs:
   build:
-    permissions:
-      actions: read
-      attestations: read
-      checks: read
-      contents: write
-      deployments: read
-      id-token: none
-      issues: none
-      discussions: none
-      packages: read
-      pages: read
-      pull-requests: write
-      security-events: read
-      statuses: read
+    permissions: read-all
 
     runs-on: ubuntu-latest
     strategy:
@@ -42,26 +29,3 @@ jobs:
       - name: Generate Coverage Report
         run: |
           coverage report -m
-      - name: Coverage Badge
-        uses: tj-actions/coverage-badge-py@1788babcb24544eb5bbb6e0d374df5d1e54e670f # v2
-
-      - name: Verify Changed files
-        uses: tj-actions/verify-changed-files@a1c6acee9df209257a246f2cc6ae8cb6581c1edf # v20
-        id: verify-changed-files
-        with:
-          files: coverage.svg
-
-      - name: Commit files
-        if: steps.verify-changed-files.outputs.files_changed == 'true'
-        run: |
-          git config --local user.email "github-actions[bot]@users.noreply.github.com"
-          git config --local user.name "github-actions[bot]"
-          git add coverage.svg
-          git commit -m "Updated coverage.svg"
-
-      - name: Push changes
-        if: steps.verify-changed-files.outputs.files_changed == 'true'
-        uses: ad-m/github-push-action@master
-        with:
-          github_token: ${{ secrets.github_token }}
-          branch: ${{ github.ref }}
