+++ .github/workflows/python-tests-coverage.yml
@@ -1,6 +1,10 @@
 name: Build tests coverage badge
 
-on: [push]
+on:
+  push:
+    branches:
+      - main
+      - dev
 
 jobs:
   build:
