### 2025-09-20 DataverseLabs/pyinterpolate#484 @SimonMolinsky
Hi @crhea93

I've added both badges to the README, code coverage is estimated using Github Actions.
