### 2025-09-23 7154ee0
[dev] inplace param added to crs transformation of Blocks object (+ declaration of new object's deepcopy)
