### 2025-09-30 DataverseLabs/pyinterpolate#488 @crhea93
Gaussian Processes with Kriging
Nope! This is simply an interesting extension that I think would be useful to add.

**Describe the solution you'd like**
It would be convenient if a Gaussian Process could be used as a theoretical model.
This could be an option for the `models_group` argument of the `build_theoretical_variogram` method.

**Additional context**
A useful paper: https://arxiv.org/html/2408.02331v1
Another paper: https://pure.strath.ac.uk/ws/portalfiles/portal/171374182/Suryasentana_Sheil_OSIG2023_Demystifying_the_connections_between_Gaussian_Process_Regression_and_Kriging.pdf
Implementation: https://scikit-learn.org/stable/modules/gaussian_process.html

This is related to the PyOpenSci submission https://github.com/pyOpenSci/software-submission/issues/250
