### 2025-10-01 DataverseLabs/pyinterpolate#488 @SimonMolinsky
Thank you, this is great idea! I will introduce this method in the next release :)
