### 2025-10-02 DataverseLabs/pyinterpolate#488 @crhea93
Fantastic!
