### 2025-10-11 e1a9758
Updated sill descriptions in documentation, updated spatial dependency index calculations
