### 2025-10-13 991fe2f
Use URLs for images and badges in README.md
