### 2025-10-16 DataverseLabs/pyinterpolate#494 @martinfleis
link to nonexistent CONTRIBUTION.md
Documentation contains a link to a file that gives 404.

From https://pyinterpolate.readthedocs.io/en/stable/contributor/development/tests_and_contribution.html to https://github.com/DataverseLabs/pyinterpolate/blob/main/CONTRIBUTION.md. It should lead to https://github.com/DataverseLabs/pyinterpolate/blob/main/CONTRIBUTING.md

xref https://github.com/pyOpenSci/software-submission/issues/250
