### 2025-10-16 DataverseLabs/pyinterpolate#495 @martinfleis
[project.urls] not in pyproject
pyproject could contain `[project.urls]` section pointing to documentation and GitHub.

xref https://github.com/pyOpenSci/software-submission/issues/250
