### 2025-10-16 DataverseLabs/pyinterpolate#496 @martinfleis
Missing examples
Some of the functions are missing Examples, like semivariograms. pyOpenSci assumes all are present.

xref https://github.com/pyOpenSci/software-submission/issues/250
