### 2025-10-16 #250 @martinfleis
- [x] As the reviewer, I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience in the README file.
- [x] **Short quickstart tutorials** demonstrating significant functionality that successfully runs locally.
- [x] **Metadata** including author(s), author e-mail(s), a URL, and any other relevant metadata, for example, in a `pyproject.toml` file or elsewhere.

*NOTE: If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be wider than high. A badge for pyOpenSci peer review will be provided when the package is accepted.*

The package structure should follow the general community best practices. In general, please consider whether:

---

pyInterpolate is a great package! It is even part of my teaching curriculum so I was happy to review it properly and provide additional feedback, partially based on that I got in class. @SimonMolinsky included all my suggestions in the current dev version, so I am very happy to suggest acceptance. Good job!
