### 2025-10-16 DataverseLabs/pyinterpolate#497 @martinfleis
Citation information
The citation information linking to the JOSS paper should be available in readme and you should ideally have a citation.cff file as well, pointing to the same.

edit: I see the badge but an explicit section with bibtex item would be nice.

xref https://github.com/pyOpenSci/software-submission/issues/250
