### 2025-10-16 DataverseLabs/pyinterpolate#498 @martinfleis
Missing badges
The pyOpenSci guidelines suggest having these badges in Radme, which are missing at the moment.

- Python versions supported,
- Current package version (on PyPI / Conda).

xref https://github.com/pyOpenSci/software-submission/issues/250
