### 2025-10-16 DataverseLabs/pyinterpolate#494 @martinfleis
The same is linked from readme.
