### 2025-10-16 DataverseLabs/pyinterpolate#499 @martinfleis
Inconsistency of API with the ecosystem
I have one comment on the API as part of the pyOpenSci review, something I've been wondering about for a while.

When passing the data composed of location (either as geometry or coords) and value, this is expected to be a single object with columns in a specific order. I have to admit I find it quite strange, as typically you are passing these as two different objects (pysal style), or if as one, then you are specifying which column is what (seaborn style). What is the logic behind this decision?

Since 1.0 it is actually a bit better as you can directly work with geopandas objects, while before you had to have a single array with coordinates and values linked together (that is super confusing in my mental model).

Have you ever considered splitting `ds` to two keywords, something like `geometry, y` that would avoid those occasionally awkward concatenations and make all of this more explicit?

https://github.com/pyOpenSci/software-submission/issues/250
