### 2025-10-17 DataverseLabs/pyinterpolate#499 @SimonMolinsky
I think (but I'm unsure) that this came from my background - in the past it had more to do with time series modeling than geospatial, where you have time indexes and observed values. In this way, coordinates are indexes, always the first two columns, and the last column is observation.

But what you are proposing makes sense, and I will introduce this enhancement. I will make `ds` optional and add two additional keywords representing geometry and observed values. This won't break anything at this stage.

Thanks for pointing this out!
