### 2025-10-17 DataverseLabs/pyinterpolate#499 @martinfleis
Cool! A similar situation happens with the objects that are returned but that is much harder to change as it is a breaking change. So I'll hold my horses on that :)
