### 2025-10-19 DataverseLabs/pyinterpolate#499 @SimonMolinsky
Class and functions to check:

- [x] Blocks (core)
- [x] PointSupport (core)
- [x] interpolate_points (core)
- [x] interpolate_points_dask (core)
- [x] validate_kriging (evaluate)
- [x] inverse_distance_weighting (idw)
- [x] IndicatorKriging (kriging)
- [x] ordinary_kriging (kriging)
- [x] ordinary_kriging_from_cov (kriging)
- [x] simple_kriging (kriging)
- [x] UniversalKriging (kriging)
- [x] calculate_average_p2b_semivariance (semivariogram)
- [x] DirectionalVariogram (semivariogram)
- [x] ExperimentalVariogram (semivariogram)
- [x] build_experimental_variogram (semivariogram)
- [x] VariogramCloud (semivariogram)
- [x] directional_weighted_semivariance (semivariogram)
- [x] omnidirectional_variogram (semivariogram)
- [x] omnidirectional_semivariogram_cloud (semivariogram)
- [x] calculate_covariance (semivariogram)
- [x] directional_covariance (semivariogram)
- [x] omnidirectional_covariance (semivariogram)
- [x] calculate_semivariance (semivariogram)
- [x] directional_semivariance (semivariogram)
- [x] directional_semivariance_cloud (semivariogram)
- [x] omnidirectional_semivariance (semivariogram)
- [x] point_cloud_semivariance (semivariogram)
- [x] code_indicators (semivariogram)
- [x] IndicatorVariogramData (semivariogram)
- [x] ExperimentalIndicatorVariogram (semivariogram)
- [x] interpolate_raster (viz)

And last but not least:

- [ ] Update tutorials
