### 2025-10-21 8906833
[enh] Blocks object may use separate geometry and values as input parameters + tests
