### 2025-10-23 DataverseLabs/pyinterpolate#499 @SimonMolinsky
@martinfleis I've started working on this issue, but I want to be sure if I understood you correctly.

As an example, I will show you what I have done with the `calculate_semivariance()` function:

```python

def calculate_semivariance(ds: Union[ArrayLike, VariogramPoints] = None,
                           values: ArrayLike = None,
                           geometries: ArrayLike = None,
                           step_size: float = None,
                           max_range: float = None,
                           direction: float = None,
                           tolerance: float = None,
                           custom_bins: Union[ArrayLike, Any] = None,
                           custom_weights: ArrayLike = None) -> np.ndarray:
    """
    Calculates experimental semivariance.

    Parameters
    ----------
    ds : ArrayLike, optional
        ``[x, y, value]``

    values : ArrayLike, optional
        Aggregated values of each block. Optional parameter, if not
        given then ``ds`` must be provided.

    geometries : ArrayLike, optional
        Array or similar structure with geometries. It must have the same
        length as ``values``. Optional parameter, if not given then ``ds``
        must be provided. Those must be point geometries!

    ...
    """
    ...

```

**Is this what you meant?** Right now user might pass `values` and `geometries` as parameters, and here is the sample test where I use this notation:

```python
def test_omnidirectional_semivariogram_values_geometry_inputs():
    import geopandas as gpd

    REFERENCE_INPUT = np.array([
        [0, 0, 8],
        [1, 0, 6],
        ...
    ])
    STEP_SIZE = 1
    MAX_RANGE = 4

    GEOMETRY = gpd.points_from_xy(
        x=REFERENCE_INPUT[:, 0], y=REFERENCE_INPUT[:, 1]
    )

    semivariance = calculate_semivariance(
        ds=None,
        values=REFERENCE_INPUT[:, -1],
        geometries=GEOMETRY,
        step_size=STEP_SIZE,
        max_range=MAX_RANGE
    )

...

```
