### 2025-10-24 DataverseLabs/pyinterpolate#499 @martinfleis
Yes!
