### 2025-10-25 6047084
updated calculate_semivariance() function
