### 2025-10-25 45ff9e8
interpolate_points() - introduction of new parameters
