### 2025-10-25 0b0c8ad
interpolate_points() function - basic tests
