### 2025-10-25 9f625d2
validate_kriging() function - changed parameters
