### 2025-10-26 1a43b5b
validate_kriging - values, geometries, and progress_bar parameters; tests
