### 2025-10-26 41c1d18
idw - known_values and known_geometries parameters, tests
