+++ src/pyinterpolate/idw/idw.py
@@ -8,16 +8,19 @@ Authors
 TODO
 - eval with IDW
 """
-from typing import Iterable
+from numpy.typing import ArrayLike
 
 import numpy as np
 
 from pyinterpolate.core.data_models.points import VariogramPoints
 from pyinterpolate.distance.point import point_distance
+from transform.geo import geometry_and_values_array
 
 
-def inverse_distance_weighting(known_locations: np.ndarray,
-                               unknown_location: Iterable,
+def inverse_distance_weighting(unknown_location: ArrayLike,
+                               known_locations: ArrayLike = None,
+                               known_values: ArrayLike = None,
+                               known_geometries: ArrayLike = None,
                                no_neighbors=-1,
                                power=2.) -> float:
     """
@@ -26,11 +29,6 @@ def inverse_distance_weighting(known_locations: np.ndarray,
 
     Parameters
     ----------
-    known_locations : numpy array
-        The MxN array, where **M** is a number of rows (points) and **N**
-        is the number of columns, where the last column represents a value
-        observed in a known point. (It could be **(N-1)**-dimensional data).
-
     unknown_location : Iterable
         Array or list with coordinates of the unknown point.
         Its length is N-1 (number of dimensions). The unknown
@@ -38,6 +36,18 @@ def inverse_distance_weighting(known_locations: np.ndarray,
         parameter `shape`, if not, then new dimension
         is added once - vector of points ``[x, y]``
         becomes ``[[x, y]]`` for 2-dimensional data.
+
+    known_locations : numpy array, optional
+        The known locations: ``[x, y, value]``.
+
+    known_values : ArrayLike, optional
+        Observation in the i-th geometry (from ``known_geometries``). Optional
+        parameter, if not given then ``known_locations`` must be provided.
+
+    known_geometries : ArrayLike, optional
+        Array or similar structure with geometries. It must have the same
+        length as ``known_values``. Optional parameter, if not given then
+        ``known_locations`` must be provided. Point type geometry.
 
     no_neighbors : int, default = -1
         If default value **(-1)** then all known points will be used to
@@ -67,6 +77,13 @@ def inverse_distance_weighting(known_locations: np.ndarray,
     # Check power parameter
     if power < 0:
         raise ValueError('Power cannot be smaller than 0')
+
+    # Get known locations
+    if known_locations is None:
+        known_locations = geometry_and_values_array(
+            geometry=known_geometries,
+            values=known_values
+        )
 
     # Check known points parameter
     # Check if known locations are in the right format
