### 2025-10-26 1c818c6
IndicatorKriging - new parameters and test
