+++ tests/test_kriging/test_indicator.py
@@ -66,3 +66,22 @@ def test_omni():
 
     imaps = ikriging.get_indicator_maps()
     assert isinstance(imaps, Dict)
+
+
+def test_omni_separate_values_and_geometry():
+    ikriging = IndicatorKriging(
+        indicator_variograms=VARIOGRAMS,
+        unknown_locations=TEST[:, :-1],
+        known_values=TRAIN[:, -1],
+        known_geometries=TRAIN[:, :-1],
+        kriging_type='ok',
+        no_neighbors=16,
+        allow_approximate_solutions=True
+    )
+    assert isinstance(ikriging, IndicatorKriging)
+    assert isinstance(ikriging.indicator_predictions, np.ndarray)
+    assert isinstance(ikriging.expected_values, np.ndarray)
+    assert isinstance(ikriging.variances, np.ndarray)
+
+    imaps = ikriging.get_indicator_maps()
+    assert isinstance(imaps, Dict)
