### 2025-10-30 bfdd039
ordinary kriging - tests, new parameters
