+++ src/pyinterpolate/kriging/point/ordinary.py
@@ -29,8 +29,10 @@ from pyinterpolate.semivariogram.theoretical.theoretical import TheoreticalVario
 
 def ok_calc(
         theoretical_model: TheoreticalVariogram,
-        known_locations: np.ndarray,
         unknown_location: ArrayLike,
+        known_locations: ArrayLike = None,
+        known_values: ArrayLike = None,
+        known_geometries: ArrayLike = None,
         neighbors_range=None,
         no_neighbors=4,
         max_tick=5.,
@@ -46,11 +48,20 @@ def ok_calc(
     theoretical_model : TheoreticalVariogram
         Fitted theoretical variogram model.
 
+    unknown_location : Union[ArrayLike, Point]
+        Points where you want to estimate value ``(x, y) <-> (lon, lat)``.
+
     known_locations : numpy array
         Known locations: ``[x, y, value]``.
 
-    unknown_location : Union[ArrayLike, Point]
-        Points where you want to estimate value ``(x, y) <-> (lon, lat)``.
+    known_values : ArrayLike, optional
+        Observation in the i-th geometry (from ``known_geometries``). Optional
+        parameter, if not given then ``known_locations`` must be provided.
+
+    known_geometries : ArrayLike, optional
+        Array or similar structure with geometries. It must have the same
+        length as ``known_values``. Optional parameter, if not given then
+        ``known_locations`` must be provided. Point type geometry.
 
     neighbors_range : float, default=None
         The maximum distance where we search for neighbors. If ``None`` is
@@ -87,7 +98,12 @@ def ok_calc(
         Singular Matrix in the Kriging system.
     """
     # Check if known locations are in the right format
-    known_locations = VariogramPoints(known_locations).points
+    # Validate points
+    if not isinstance(known_locations, VariogramPoints):
+        known_locations = VariogramPoints(points=known_locations,
+                                          geometries=known_geometries,
+                                          values=known_values)
+        known_locations = known_locations.points
 
     # Check if unknown location is Point
     if isinstance(unknown_location, Point):
@@ -133,8 +149,10 @@ def ok_calc(
 
 def ordinary_kriging(
         theoretical_model: TheoreticalVariogram,
-        known_locations: ArrayLike,
         unknown_locations: Union[np.ndarray, Point, List, Tuple, GeoSeries, GeometryArray, ArrayLike],
+        known_locations: ArrayLike = None,
+        known_values: ArrayLike = None,
+        known_geometries: ArrayLike = None,
         neighbors_range=None,
         no_neighbors=4,
         max_tick=5.,
@@ -151,11 +169,20 @@ def ordinary_kriging(
     theoretical_model : TheoreticalVariogram
         Fitted theoretical variogram model.
 
+    unknown_locations : Union[ArrayLike, Point]
+        Points where you want to estimate value ``(x, y) <-> (lon, lat)``.
+
     known_locations : numpy array
         Known locations: ``[x, y, value]``.
 
-    unknown_locations : Union[ArrayLike, Point]
-        Points where you want to estimate value ``(x, y) <-> (lon, lat)``.
+    known_values : ArrayLike, optional
+        Observation in the i-th geometry (from ``known_geometries``). Optional
+        parameter, if not given then ``known_locations`` must be provided.
+
+    known_geometries : ArrayLike, optional
+        Array or similar structure with geometries. It must have the same
+        length as ``known_values``. Optional parameter, if not given then
+        ``known_locations`` must be provided. Point type geometry.
 
     neighbors_range : float, default=None
         The maximum distance where we search for neighbors. If ``None`` is
@@ -195,7 +222,12 @@ def ordinary_kriging(
         Singular Matrix in the Kriging system.
     """
     # Check if known locations are in the right format
-    known_locations = VariogramPoints(known_locations).points
[…]
