+++ src/pyinterpolate/transform/geo.py
@@ -6,6 +6,7 @@ import numpy as np
 import pandas as pd
 
 from shapely.geometry import Polygon, MultiPolygon, Point
+from shapely.geometry.base import BaseGeometry
 
 
 def geometry_and_values_array(geometry,
@@ -81,10 +82,10 @@ def largest_geometry(geometry: MultiPolygon) -> Polygon:
     return poly
 
 
-def join_geometry_and_values(geometry,
-                             values,
-                             geometry_column_name: str = 'geometry',
-                             values_column_name: str = 'values') -> gpd.GeoDataFrame:
+def join_any_geometry_and_values(geometry,
+                                 values,
+                                 geometry_column_name: str = 'geometry',
+                                 values_column_name: str = 'values') -> gpd.GeoDataFrame:
     """
     Function creates single object from geometries and aggregated values.
 
@@ -122,8 +123,65 @@ def join_geometry_and_values(geometry,
         gdf = geometry.copy(deep=True)
         gdf[values_column_name] = values
     else:
+        g0 = geometry[0]
+        if isinstance(g0, BaseGeometry):
+            gdf = gpd.GeoDataFrame(
+                values, columns=[values_column_name], geometry=geometry
+            )
+            gdf.columns = [values_column_name, geometry_column_name]
+        else:
+            raise TypeError('Passed geometry must be basic geometry types'
+                            ' supported by shapely (Point, MultiPoint,'
+                            'LineString, MultiLineString, Polygon, '
+                            'MultiPolygon)')
+
+    return gdf
+
+
+def join_point_geometry_and_values(geometry,
+                                   values,
+                                   geometry_column_name: str = 'geometry',
+                                   values_column_name: str = 'values') -> gpd.GeoDataFrame:
+    """
+    Function creates single object from geometries and aggregated values.
+
+    Parameters
+    ----------
+    geometry : ArrayLike
+
+    values : ArrayLike
+
+    geometry_column_name : str, default = 'geometry'
+
+    values_column_name : str, default = 'value'
+
+    Returns
+    -------
+    : gpd.GeoDataFrame
+    """
+
+    if len(geometry) != len(values):
+        raise ValueError(
+            'Number of geometries must be equal to number of values'
+        )
+
+    if isinstance(values, pd.Series):
+        values = values.values
+    elif isinstance(values, pd.DataFrame):
+        val_column_name = values.columns[0]
+        values = values[val_column_name].values
+
+    if isinstance(geometry, pd.DataFrame):
+        geom_column_name = geometry.columns[0]
+        geometry = geometry[geom_column_name].values
+
+    if isinstance(geometry, gpd.GeoDataFrame):
+        gdf = geometry.copy(deep=True)
+        gdf[values_column_name] = values
+    else:
+        points = [Point(p) for p in geometry]
         gdf = gpd.GeoDataFrame(
-            values, columns=[values_column_name], geometry=geometry
+            values, columns=[values_column_name], geometry=points
         )
         gdf.columns = [values_column_name, geometry_column_name]
 
