### 2025-10-30 87e9669
simple kriging - new parameters, new tests
