+++ CHANGELOG.rst
@@ -15,6 +15,8 @@ Changes - from version >= 1.x
 * [enhancement] `IndicatorKriging` class has the new parameters `known_values` and `known_geometries`, that might be provided instead of the `known_locations` parameter
 * [enhancement] `ordinary_kriging()` function has the new parameters `known_values` and `known_geometries`, that might be provided instead of the `known_locations` parameter
 * [enhancement] `simple_kriging()` function has the new parameters `known_values` and `known_geometries`, that might be provided instead of the `known_locations` parameter
+* [enhancement] `UniversalKriging` class has the new parameters `known_values` and `known_geometries`, that might be provided instead of the `known_locations` parameter
+
 
 2025-10-11
 ----------
