### 2025-11-01 f6d8451
make inblock_semivariance() private function, it should't be called by users
