+++ CHANGELOG.rst
@@ -16,7 +16,8 @@ Changes - from version >= 1.x
 * [enhancement] `ordinary_kriging()` function has the new parameters `known_values` and `known_geometries`, that might be provided instead of the `known_locations` parameter
 * [enhancement] `simple_kriging()` function has the new parameters `known_values` and `known_geometries`, that might be provided instead of the `known_locations` parameter
 * [enhancement] `UniversalKriging` class has the new parameters `known_values` and `known_geometries`, that might be provided instead of the `known_locations` parameter
-
+* [api change] `inblock_semivariance()` function is now private: `_inblock_semivariance()`
+* 
 
 2025-10-11
 ----------
