### 2025-11-02 bb2d0e2
DirectionalVariogram - new parameters, new tests
