### 2025-11-02 a1714b0
ExperimentalVariogram - new parameters and new tests
