### 2025-11-02 b43fa1c
build_experimental_variogram() - new tests, new parameters
