### 2025-11-02 b5131bd
VariogramCloud - new parameters and one test
