### 2025-11-02 d65961b
[api change] directional_weighted_semivariance is no longer public function
