### 2025-11-03 55246b1
multiple api changes and tests
