### 2025-11-04 8836cb4
code_indicator() function has new parameters and tests
