### 2025-11-04 2d0c94d
IndicatorVariogramData and ExperimentalIndicatorVariogram classes have new parameters and tests
