+++ CHANGELOG.rst
@@ -35,6 +35,7 @@ Changes - from version >= 1.x
 * [api change] `omnidirectional_semivariance()` function is now private: `_omnidirectional_semivariance()`
 * [enhancement] `point_cloud_semivariance()` function has the new parameters `values` and `geometries`, that might be provided instead of `ds` parameter
 * [enhancement] `code_indicator()` function has the new parameters `values` and `geometries`, that might be provided instead of `ds` parameter
+* [enhancement] `ExperimentalIndicatorVariogram` and `IndicatorVariogramData` classes have the new parameters `values` and `geometries`, that might be provided instead of `ds` parameter
 
 2025-10-11
 ----------
