### 2025-11-04 10e655a
interpolate_raster - changed parameters naming and new tests
