### 2025-11-06 d29345a
PointSupport - separate geometry and values
