+++ docs/build/html/_sources/usage/tutorials/functional/1-3-spatial-dependency-index.ipynb.txt
@@ -90,8 +90,8 @@
   {
    "metadata": {
     "ExecuteTime": {
-     "end_time": "2025-04-23T18:16:10.015671Z",
-     "start_time": "2025-04-23T18:16:06.038465Z"
+     "end_time": "2025-11-07T11:30:04.358783Z",
+     "start_time": "2025-11-07T11:30:01.112010Z"
     }
    },
    "cell_type": "code",
@@ -107,8 +107,8 @@
   {
    "metadata": {
     "ExecuteTime": {
-     "end_time": "2025-04-23T18:16:10.038172Z",
-     "start_time": "2025-04-23T18:16:10.033735Z"
+     "end_time": "2025-11-07T11:30:07.551854Z",
+     "start_time": "2025-11-07T11:30:07.548160Z"
     }
    },
    "cell_type": "code",
@@ -133,8 +133,8 @@
   {
    "metadata": {
     "ExecuteTime": {
-     "end_time": "2025-04-23T18:16:11.191899Z",
-     "start_time": "2025-04-23T18:16:11.145321Z"
+     "end_time": "2025-11-07T11:30:09.216854Z",
+     "start_time": "2025-11-07T11:30:09.196873Z"
     }
    },
    "cell_type": "code",
@@ -248,8 +248,8 @@
   {
    "metadata": {
     "ExecuteTime": {
-     "end_time": "2025-04-23T18:16:55.248117Z",
-     "start_time": "2025-04-23T18:16:52.956964Z"
+     "end_time": "2025-11-07T11:31:10.172971Z",
+     "start_time": "2025-11-07T11:31:08.028342Z"
     }
    },
    "cell_type": "code",
@@ -261,7 +261,10 @@
     "    arr = ds.values\n",
     "\n",
     "    # Get experimental variogram\n",
-    "    exp_var = ptp.build_experimental_variogram(arr, step_size=STEP_SIZE, max_range=MAX_RANGE)\n",
+    "    exp_var = ptp.build_experimental_variogram(values=ds[element],\n",
+    "                                               geometries=ds[['x', 'y']],\n",
+    "                                               step_size=STEP_SIZE,\n",
+    "                                               max_range=MAX_RANGE)\n",
     "\n",
     "    # Find optimal theoretical model\n",
     "    optimal_model = ptp.TheoreticalVariogram()\n",
@@ -279,7 +282,7 @@
      "name": "stdout",
      "output_type": "stream",
      "text": [
-      "Optimal model for element cadmium has 43.96% of nugget to sill ratio. Spatial Dependence is moderate.\n"
+      "Optimal model for element cadmium has 30.54% of nugget to sill ratio. Spatial Dependence is moderate.\n"
      ]
     },
     {
@@ -299,7 +302,7 @@
       "\n",
       "---\n",
       "\n",
-      "Optimal model for element copper has 26.08% of nugget to sill ratio. Spatial Dependence is moderate.\n"
+      "Optimal model for element copper has 20.69% of nugget to sill ratio. Spatial Dependence is strong.\n"
      ]
     },
     {
@@ -319,7 +322,7 @@
       "\n",
       "---\n",
       "\n",
-      "Optimal model for element zinc has 1.09% of nugget to sill ratio. Spatial Dependence is strong.\n"
+      "Optimal model for element zinc has 1.08% of nugget to sill ratio. Spatial Dependence is strong.\n"
      ]
     },
     {
@@ -339,7 +342,7 @@
       "\n",
       "---\n",
       "\n",
-      "Optimal model for element lead has 14.34% of nugget to sill ratio. Spatial Dependence is strong.\n"
+      "Optimal model for element lead has 12.54% of nugget to sill ratio. Spatial Dependence is strong.\n"
      ]
     },
     {
[…]
