+++ docs/build/html/_sources/usage/tutorials/functional/4-2-poisson-kriging-centroid-based.ipynb.txt
@@ -49,8 +49,8 @@
    "metadata": {
     "collapsed": true,
     "ExecuteTime": {
-     "end_time": "2025-07-11T07:02:36.372840Z",
-     "start_time": "2025-07-11T07:02:33.057175Z"
+     "end_time": "2025-11-07T19:30:23.797631Z",
+     "start_time": "2025-11-07T19:30:21.015844Z"
     }
    },
    "cell_type": "code",
@@ -70,8 +70,8 @@
   {
    "metadata": {
     "ExecuteTime": {
-     "end_time": "2025-07-11T07:02:36.382189Z",
-     "start_time": "2025-07-11T07:02:36.379173Z"
+     "end_time": "2025-11-07T19:30:23.958996Z",
+     "start_time": "2025-11-07T19:30:23.955984Z"
     }
    },
    "cell_type": "code",
@@ -92,8 +92,8 @@
   {
    "metadata": {
     "ExecuteTime": {
-     "end_time": "2025-07-11T07:02:36.837310Z",
-     "start_time": "2025-07-11T07:02:36.565125Z"
+     "end_time": "2025-11-07T19:30:24.273049Z",
+     "start_time": "2025-11-07T19:30:23.990200Z"
     }
    },
    "cell_type": "code",
@@ -112,8 +112,8 @@
   {
    "metadata": {
     "ExecuteTime": {
-     "end_time": "2025-07-11T07:02:37.217414Z",
-     "start_time": "2025-07-11T07:02:36.854403Z"
+     "end_time": "2025-11-07T19:30:24.584955Z",
+     "start_time": "2025-11-07T19:30:24.282470Z"
     }
    },
    "cell_type": "code",
@@ -133,8 +133,8 @@
   {
    "metadata": {
     "ExecuteTime": {
-     "end_time": "2025-07-11T07:02:38.159177Z",
-     "start_time": "2025-07-11T07:02:37.227490Z"
+     "end_time": "2025-11-07T19:30:26.456081Z",
+     "start_time": "2025-11-07T19:30:25.541997Z"
     }
    },
    "cell_type": "code",
@@ -190,8 +190,8 @@
   {
    "metadata": {
     "ExecuteTime": {
-     "end_time": "2025-07-11T07:02:38.171807Z",
[…]
