### 2025-11-08 DataverseLabs/pyinterpolate#499 @SimonMolinsky
Ok, I've done it. Issue will stay open till I update docstrings with the examples using the new parameters, but it is a part of the issue #496
