### 2025-11-09 0daba1c
Added supported python version and package version badges
