### 2025-11-09 3f19ee5
CITATION.cff, citation section in readme has been moved up, bibtex citation has been added
