+++ CHANGELOG.rst
@@ -41,6 +41,7 @@ Changes - from version >= 1.x
 * [docs] added missing badges (Python version and package version)
 * [docs] `CITATION.cff`, citation section in readme has been moved up, bibtex citation has been added
 * [docs] corrected broken links to `CONTRIBUTING.md` file in readme and docs
+* [setup] links to documentation have been added to `pyproject.toml`
 * # todo [docs] docstring examples - all user-facing API endpoints has examples in docstrings, using new input types (values | geometries)
 
 2025-10-11
