### 2025-11-10 38eeba7
Updated docstring (example) - `PointSupportDistance`
