### 2025-11-11 d04efc5
block_filter module - updated examples in docstrings
