### 2025-11-11 cd083c6
updated docstrings in interpolate module (examples)
