### 2025-11-12 9e7d982
Updated docs in `pyinterpolate.distance.block.calc_block_to_block_distance`, `validate_kriging`, `metrics` module functions
