### 2025-11-13 c12cb5b
idw | centroid_poisson_kriging docs examples
