### 2025-11-15 156a9ae
Kriging documentation has been updated
