### 2025-11-16 71d4d86
updated docs, deviation api changes, deviation tests, new properties
