### 2025-11-18 570d1f1
Updated and debugged Deviation class
