### 2025-11-18 3da08f5
docstring example for deviation class
