### 2025-11-18 4db4790
docstring examples for `AggregatedVariogram` and `regularize()`
