+++ CHANGELOG.rst
@@ -49,7 +49,8 @@ Changes - from version >= 1.x
 * [api change] `Deviation` class parameters have been renamed to clearly point what is passed as the input
 * [tests] `Deviation` class tests and debug (method selection)
 * [docs] docstring example for `Deviation`
-* 
+* [docs] docstring examples for `AggregatedVariogram` and `regularize()`
+*
 
 2025-10-11
 ----------
