### 2025-11-18 65a0862
docstring examples for `ExperimentalVariogram` and `build_experimental_variogram()`, used parameters `values` and `geometries`
