### 2025-11-18 a4cc2a4
calculate_semivariance() example
