### 2025-11-19 7f3f996
Updated calculate_covariance() docstring
