### 2025-11-19 b80949e
Directional variogram example
