### 2025-11-23 74a9ad0
`ExperimentalVariogram` class removes lags with NaN values (in covariance and semivariance); Example in `interpolate_raster()` function
