+++ src/pyinterpolate/semivariogram/deconvolution/regularize.py
@@ -177,8 +177,6 @@ class Deconvolution:
     * Mean Bias: None
     * Mean RMSE: 16.524700893642784
     * Error-lag weighting method: closest
-
-
     +----------+--------------------+--------------------+---------------------+
     |   lag    |    theoretical     |    experimental    |   bias (real-yhat)  |
     +----------+--------------------+--------------------+---------------------+
@@ -197,6 +195,7 @@ class Deconvolution:
     | 260000.0 | 157.72180532432975 | 125.00372185725449 |  -32.71808346707526 |
     | 280000.0 | 157.72180532432975 | 122.2260745900828  |  -35.49573073424695 |
     +----------+--------------------+--------------------+---------------------+
+
     >>> dcv.export_model_to_json('results.csv')
     """
 
