### 2025-11-23 DataverseLabs/pyinterpolate#496 @SimonMolinsky
I've added examples to all user-facing functions
