### 2025-11-23 DataverseLabs/pyinterpolate#499 @SimonMolinsky
I've added examples to docstrings using new API structure, also I changed tutorials to use this new version where geometry and values are separated.
