### 2025-11-23 cdd103d
pre-bump version updates
