+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
 Changes - from version >= 1.x
 =============================
 
-2025-11-1
+2025-11-23
 ----------
 
 **version 1.2.0**
