### 2025-11-23 da7f2f3
remove Python 3.14 from supported versions (pyogrio installation problems)
