### 2025-12-01 #250 @banesullivan
Thank you for the thorough reviews @martinfleis and @crhea93 and for opening and linking several issues to this review!! @SimonMolinsky, excellent work on pyinterpolate, this has come together fantastically!!

*Aside: apologies for my delay!! Life happened and before I knew it a few weeks went by. So sorry about that!*

A few quick questions before I post the review acceptance and update the metadata in the original post at the top of this thread:

- Is the submitted Version 1.0.1 the latest incorporating any/all changes requested from this review? If not, @SimonMolinsky, would you be able to post a release and update me with the version to tag with this review/acceptance?
- Do you have a Zendo version and DOI for this? I see this has already been reviewed/accepted to JOSS and the JOSS DOI which is sufficient.

Otherwise, I'm excited accept pyinterpolate and I'll start double checking to finalize the review.
