### 2025-12-02 #250 @SimonMolinsky
@banesullivan

No apologies required, but thank you for that!

Going back to the package and review:

- I've published version 1.2.0 here: [https://github.com/DataverseLabs/pyinterpolate/releases/tag/v1.2.0](https://github.com/DataverseLabs/pyinterpolate/releases/tag/v1.2.0)
- DOI for JOSS paper is: https://doi.org/10.21105/joss.02869
- DOI for this version on Zenodo (1.2.0) is: https://doi.org/10.5281/zenodo.17691515
