### 2025-12-03 #250 @banesullivan
Fantastic! I've updated the metadata in the top level post here to reflect those details.

🎉 With that, pyinterpolate has been approved by pyOpenSci!

- [x] Add the badge for pyOpenSci peer-review to the README.md of pyinterpolate. The badge should be `[![pyOpenSci Peer-Reviewed](https://pyopensci.org/badges/peer-reviewed.svg)](https://github.com/pyOpenSci/software-review/issues/250)`.

🎉 Congratulations! You are now published with pyOpenSci in addition to JOSS! 🎉

- [x] Change the status tag of the issue to `6/pyOS-approved 🚀🚀🚀`.  (You can remove all other labels; If the the package is moving on to JOSS then add a JOSS label)
- [x] Update the header of this issue with the date accepted, the version accepted, and the accepted version DOI (from Zenodo) to the header of this issue.
- [x] Invite the package maintainer(s) and both reviewers to the pyOpenSci Slack.

---
