### 2025-12-03 #250 @banesullivan
@SimonMolinsky, I like to invite you to write a blog post (totally optional) on pyinterpolate for us to promote your work! if you are interested - here are a few examples of other blog posts:

- [pandera](https://www.pyopensci.org/blog/pandera-python-pandas-dataframe-validation)
- [movingpandas](https://www.pyopensci.org/blog/movingpandas-movement-data-analysis)

and [here](https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_posts/2019-11-18-pandera-dataframe-validation.md) is a markdown example that you could use as a guide when creating your post.

it can even be a tutorial like post that highlights what your package does. Then we can share it with people to get the word out about pyinterpolate.

If you are too busy for this no worries. But if you have time - we'd love to spread the word about your package!
