### 2025-12-21 622f2cf
[enhancement] Handling missing geometries in Blocks and changed deafult name of values column in Blocks to avoid collisions with the deafult values column in PointSupport
