+++ src/pyinterpolate/core/data_models/blocks.py
@@ -17,8 +17,7 @@ from numpy.typing import ArrayLike
 import geopandas as gpd
 import numpy as np
 import pandas as pd
-
-from shapely.geometry import Polygon
+from shapely import MultiPolygon
 
 from pyinterpolate.distance.angular import calc_angles
 from pyinterpolate.distance.point import point_distance
@@ -219,7 +218,7 @@ class Blocks:
             self.ds = ds.copy(deep=True)
         else:
             if value_column_name is None:
-                value_column_name = 'values'
+                value_column_name = 'block_value'
             self.ds = join_any_geometry_and_values(geometry=geometries,
                                                    values=values,
                                                    values_column_name=value_column_name)
@@ -716,6 +715,6 @@ class Blocks:
             present.
         """
         ds = geometries.apply(
-            lambda x: x if isinstance(x, Polygon) else largest_geometry(x)
+            lambda x: largest_geometry(x) if isinstance(x, MultiPolygon) else x
         )
         return ds
