### 2025-12-21 8d1521b
[debug] create default index column when geometries and values are passed as arrays to Blocks
