+++ CHANGELOG.rst
@@ -1,6 +1,15 @@
 Changes - from version >= 1.x
 =============================
 
+2025-12-21
+----------
+
+**version 1.2.1**
+
+* [enhancement] `Blocks` can handle missing geometries (geometry as `None` or `NaN`)
+* [debug] `Blocks` default value column is `block_values`, and `PointSupport` default value column is `values` (to avoiding naming collisions during the join operation)
+* [debug] `Blocks` object default index is created when it is not provided by the user
+
 2025-11-23
 ----------
 
