+++ docs/source/index.rst
@@ -10,17 +10,16 @@ Pyinterpolate
   :width: 400
   :alt: Pyinterpolate Logo
 
-**version 1.2.0**
+**version 1.2.1**
 -----------------
 
 .. note::
-   The last documentation update: *2025-11-23*
+   The last documentation update: *2025-12-26*
 
 Important notice
 ................
 
 The package was updated to version 1.0 in June 2025. There are breaking API changes, so please, refer to the changelog, to know more about the changes.
-Right now, the package in version 1.0.0 is in the **beta** stage, which means that it is stable but be careful with the production use. There might be some minor bugs, and large swaths of code are not optimized yet. If you find any bugs, please report them in the `Github issue tracker <https://github.com/DataverseLabs/pyinterpolate/issues>`_.
 
 Citation
 --------
