### 2026-01-18 e1fe93d
[dev] LSA only ordinary kriging and experiments
