### 2026-01-18 e18d0ff
Update CHANGELOG.rst
